Fixed minor issue in notebooks, new functions
The notebooks did not used the filtered data in case that the dataset
was postprocessed.

The signal function now contains a new function to get sliding window
data. Furthermore, the function 'extract_trials2' provides an interface
to get up to three different classes out of trial data. However, this
function will have to be merged with the already existing
extract_trials.